+++ server/standalone/dbal/model.go
@@ -2,20 +2,62 @@ package dbal
 
 import (
 	"errors"
+	"github.com/jackc/pgtype"
 	"strconv"
+	"time"
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/jackc/pgx/v4"
 	"github.com/owkin/orchestrator/lib/asset"
 	"github.com/owkin/orchestrator/lib/common"
 	orcerrors "github.com/owkin/orchestrator/lib/errors"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
-func (d *DBAL) GetModel(key string) (*asset.Model, error) {
-	row := d.tx.QueryRow(d.ctx, `select asset from "models" where id=$1 and channel=$2`, key, d.channel)
+type sqlModel struct {
+	Key            string
+	Category       asset.ModelCategory
+	ComputeTaskKey string
+	Address        pgtype.Text
+	Checksum       pgtype.Text
+	Permissions    asset.Permissions
+	Owner          string
+	CreationDate   time.Time
+}
 
-	model := new(asset.Model)
-	err := row.Scan(model)
+func (m *sqlModel) toModel() *asset.Model {
+	model := &asset.Model{
+		Key:            m.Key,
+		Category:       m.Category,
+		ComputeTaskKey: m.ComputeTaskKey,
+		Permissions:    &m.Permissions,
+		Owner:          m.Owner,
+		CreationDate:   timestamppb.New(m.CreationDate),
+	}
+
+	if m.Address.Status == pgtype.Present {
+		model.Address = &asset.Addressable{
+			Checksum:       m.Checksum.String,
+			StorageAddress: m.Address.String,
+		}
+	}
+
+	return model
+}
+
+func (d *DBAL) GetModel(key string) (*asset.Model, error) {
+	stmt := getStatementBuilder().
+		Select("key", "compute_task_key", "category", "address", "checksum", "permissions", "owner", "creation_date").
+		From("expanded_models").
+		Where(sq.Eq{"channel": d.channel, "key": key})
+
+	row, err := d.queryRow(stmt)
+	if err != nil {
+		return nil, err
+	}
+
+	m := new(sqlModel)
+	err = row.Scan(&m.Key, &m.ComputeTaskKey, &m.Category, &m.Address, &m.Checksum, &m.Permissions, &m.Owner, &m.CreationDate)
 
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
@@ -24,7 +66,7 @@ func (d *DBAL) GetModel(key string) (*asset.Model, error) {
 		return nil, err
 	}
 
-	return model, nil
+	return m.toModel(), nil
 }
 
 func (d *DBAL) QueryModels(c asset.ModelCategory, p *common.Pagination) ([]*asset.Model, common.PaginationToken, error) {
@@ -33,17 +75,17 @@ func (d *DBAL) QueryModels(c asset.ModelCategory, p *common.Pagination) ([]*asse
 		return nil, "", err
 	}
 
-	pgDialect := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
-	stmt := pgDialect.Select("asset").
-		From("models").
+	stmt := getStatementBuilder().
+		Select("key", "compute_task_key", "category", "address", "checksum", "permissions", "owner", "creation_date").
+		From("expanded_models").
 		Where(sq.Eq{"channel": d.channel}).
-		OrderByClause("asset->>'creationDate' ASC, id").
+		OrderByClause("creation_date ASC, key").
 		Offset(uint64(offset)).
 		// Fetch page size + 1 elements to determine whether there is a next page
 		Limit(uint64(p.Size + 1))
 
 	if c != asset.ModelCategory_MODEL_UNKNOWN {
-		stmt = stmt.Where(sq.Eq{"asset->>'category'": c.String()})
+		stmt = stmt.Where(sq.Eq{"category": c.String()})
 	}
 
 	rows, err := d.query(stmt)
@@ -56,21 +98,21 @@ func (d *DBAL) QueryModels(c asset.ModelCategory, p *common.Pagination) ([]*asse
 	var count int
 
 	for rows.Next() {
-		model := new(asset.Model)
+		m := new(sqlModel)
 
-		err = rows.Scan(model)
+		err = rows.Scan(&m.Key, &m.ComputeTaskKey, &m.Category, &m.Address, &m.Checksum, &m.Permissions, &m.Owner, &m.CreationDate)
 		if err != nil {
 			return nil, "", err
 		}
 
-		models = append(models, model)
+		models = append(models, m.toModel())
 		count++
 
 		if count == int(p.Size) {
 			break
 		}
 	}
-	if err := rows.Err(); err != nil {
+	if err = rows.Err(); err != nil {
 		return nil, "", err
 	}
 
@@ -84,16 +126,29 @@ func (d *DBAL) QueryModels(c asset.ModelCategory, p *common.Pagination) ([]*asse
 }
 
 func (d *DBAL) ModelExists(key string) (bool, error) {
-	row := d.tx.QueryRow(d.ctx, `select count(id) from "models" where id=$1 and channel=$2`, key, d.channel)
+	stmt := getStatementBuilder().
+		Select("COUNT(key)").
+		From("models").
+		Where(sq.Eq{"channel": d.channel, "key": key})
+
+	row, err := d.queryRow(stmt)
+	if err != nil {
+		return false, err
+	}
 
 	var count int
-	err := row.Scan(&count)
+	err = row.Scan(&count)
 
 	return count == 1, err
 }
 
 func (d *DBAL) GetComputeTaskOutputModels(key string) ([]*asset.Model, error) {
-	rows, err := d.tx.Query(d.ctx, `select asset from "models" where compute_task_id = $1 and channel=$2`, key, d.channel)
+	stmt := getStatementBuilder().
+		Select("key", "compute_task_key", "category", "address", "checksum", "permissions", "owner", "creation_date").
+		From("expanded_models").
+		Where(sq.Eq{"channel": d.channel, "compute_task_key": key})
+
+	rows, err := d.query(stmt)
 	if err != nil {
 		return nil, err
 	}
@@ -101,15 +156,16 @@ func (d *DBAL) GetComputeTaskOutputModels(key string) ([]*asset.Model, error) {
 
 	models := []*asset.Model{}
 	for rows.Next() {
-		model := new(asset.Model)
-		err := rows.Scan(model)
+		m := new(sqlModel)
+
+		err = rows.Scan(&m.Key, &m.ComputeTaskKey, &m.Category, &m.Address, &m.Checksum, &m.Permissions, &m.Owner, &m.CreationDate)
 		if err != nil {
 			return nil, err
 		}
-		models = append(models, model)
+		models = append(models, m.toModel())
 	}
 
-	if err := rows.Err(); err != nil {
+	if err = rows.Err(); err != nil {
 		return nil, err
 	}
 
@@ -117,13 +173,68 @@ func (d *DBAL) GetComputeTaskOutputModels(key string) ([]*asset.Model, error) {
 }
 
 func (d *DBAL) AddModel(model *asset.Model) error {
-	stmt := `insert into "models" ("id", "asset", "channel", "compute_task_id") values ($1, $2, $3, $4)`
-	_, err := d.tx.Exec(d.ctx, stmt, model.GetKey(), model, d.channel, model.ComputeTaskKey)
-	return err
+	err := d.addAddressable(model.Address)
+	if err != nil {
+		return err
+	}
+
+	stmt := getStatementBuilder().
+		Insert("models").
+		Columns("key", "channel", "compute_task_key", "category", "address", "permissions", "owner", "creation_date").
+		Values(model.Key, d.channel, model.ComputeTaskKey, model.Category, model.Address.StorageAddress, model.Permissions, model.Owner, model.CreationDate.AsTime())
+
+	return d.exec(stmt)
 }
 
 func (d *DBAL) UpdateModel(model *asset.Model) error {
-	stmt := `update "models" set asset = $2, compute_task_id = $3 where id = $1 and channel = $4`
-	_, err := d.tx.Exec(d.ctx, stmt, model.GetKey(), model, model.ComputeTaskKey, d.channel)
-	return err
+	selectAddressStmt := getStatementBuilder().
+		Select("address").
+		From("models").
+		Where(sq.Eq{"channel": d.channel, "key": model.Key})
+
+	row, err := d.queryRow(selectAddressStmt)
+	if err != nil {
+		return err
+	}
+
+	var previousAddress string
+	err = row.Scan(&previousAddress)
+	if err != nil {
+		return err
+	}
+
+	updateStmt := getStatementBuilder().
+		Update("models").
+		Set("compute_task_key", model.ComputeTaskKey).
+		Set("category", model.Category).
+		Set("address", nil).
+		Set("permissions", model.Permissions).
+		Set("owner", model.Owner).
+		Where(sq.Eq{"channel": d.channel, "key": model.Key})
+
+	err = d.exec(updateStmt)
+	if err != nil {
+		return err
+	}
+
+	err = d.deleteAddressable(previousAddress)
+	if err != nil {
+		return err
+	}
+
+	if model.Address != nil {
+		err = d.addAddressable(model.Address)
+		if err != nil {
+			return err
+		}
+
+		updateAddressStmt := getStatementBuilder().
+			Update("models").
+			Set("address", model.Address.StorageAddress).
+			Where(sq.Eq{"channel": d.channel, "key": model.Key})
+
+		return d.exec(updateAddressStmt)
+	}
+
+	return nil
 }